requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
pillow>=9.2.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
tqdm>=4.64.0
tensorboard>=2.10.0
timm>=0.6.7
einops>=0.4.1
facenet-pytorch>=0.1.1
